Solved fixed sliding window problems
Goal is to solve a problem in O(n).
Given unsorted array, if it requires to pick a subarray of k based on a specific condition, create a variable that gets the window of the subarray, and go through the for loop of length of the array minus k length of the subarray and conducts a specific condition.